完成用户添加 并且 修复用户管理页面分页的问题 1.完成了用户添加的接口并且修复了接口的分页bug 2.去除冗余字段(username)改为(name)并且更新了SQL文件（jfun2019年3月26日.sql）

+++ src/main/java/cn/xiejx/jfun/dao/MenuMapper.java
@@ -13,5 +13,5 @@ import java.util.List;
  */
 @Mapper
 public interface MenuMapper extends BaseMapper<Menu> {
-    public List<Menu> getMenuByRole(@Param("roleId") Integer roleId);
+    public List<Menu> getMenuByRole(@Param("roleId") Long roleId);
 }
